requirements.txt
asyncpg>=0.29.0
langchain
langchain_text_splitters
langchain-chroma
langchain-community
langchain-huggingface
langchain-groq
langchain_core
langgraph>=0.2.0
langgraph-checkpoint-postgres
psycopg-pool
psycopg-binary
sentence-transformers
chainlit
pypdf
python-docx
docx2txt
streamlit
chromadb
pywin32
python-dotenv
tavily-python
langchain_google_genai